+++ fonctions/rules.py
@@ -1,4 +1,3 @@
-import asyncio
 import discord
 from discord.ext import commands
 import json
@@ -12,35 +11,38 @@ class Rules(commands.Cog):
 
     @commands.command()
     async def makeRules(self, ctx, *, msg):
-        id_server = ctx.message.channel.guild.id
-        with open("config.json", "r") as f:
-            config = json.load(f)
-        await ctx.message.delete()
-        rules = msg.split(" | ")
-        if str(id_server) in config:
-            config[str(id_server)]["regles"] = []
-            config[str(id_server)]["regles"] = rules
-            with open("config.json", "w") as f:
-                json.dump(config, f, indent=2)
-        else:
-            config[str(id_server)] = {}
-            config[str(id_server)]["regles"] = []
-            config[str(id_server)]["regles"] = rules
-            with open("config.json", "w") as f:
-                json.dump(config, f, indent=2)
-        i = 0
-        definedRules = ""
-        for rule in rules:
-            i += 1
-            print(i)
-            definedRules = f"{definedRules} \n {i}.{rule}"
-        embed = discord.Embed(
-            title='Discord Rules',
-            description=definedRules,
-            color=discord.Colour.dark_red()
-        )
-        embed.set_author(name='BibliBot')
-        await ctx.send(embed=embed)
+        guild_owner = ctx.message.guild.owner
+        member = ctx.message.author
+        if member == guild_owner:
+            id_server = ctx.message.channel.guild.id
+            with open("config.json", "r") as f:
+                config = json.load(f)
+            await ctx.message.delete()
+            rules = msg.split(" | ")
+            if str(id_server) in config:
+                config[str(id_server)]["regles"] = []
+                config[str(id_server)]["regles"] = rules
+                with open("config.json", "w") as f:
+                    json.dump(config, f, indent=2)
+            else:
+                config[str(id_server)] = {}
+                config[str(id_server)]["regles"] = []
+                config[str(id_server)]["regles"] = rules
+                with open("config.json", "w") as f:
+                    json.dump(config, f, indent=2)
+            i = 0
+            definedRules = ""
+            for rule in rules:
+                i += 1
+                print(i)
+                definedRules = f"{definedRules} \n {i}.{rule}"
+            embed = discord.Embed(
+                title='Discord Rules',
+                description=definedRules,
+                color=discord.Colour.dark_red()
+            )
+            embed.set_author(name='BibliBot')
+            await ctx.send(embed=embed)
 
     @commands.command()
     async def regles(self, ctx):
